refactor: move the top-level test result cluster to homeboy-extension-contract
The finale of the test subsystem breakdown. With every field dependency now in a
contract/leaf crate (via the preceding test cuts + CiContext + AppliedRefactor
already in homeboy-refactor-contract), the five top-level test result aggregates
move wholesale: TestCommandOutput, TestRunWorkflowResult, MainTestWorkflowResult,
DriftWorkflowResult, AutoFixDriftWorkflowResult.

Added homeboy-refactor-contract as a direct extension-contract dep (a pure
serde-only leaf) for AppliedRefactor. The test command/run/workflow builder
functions stay in core. Core re-exports the aggregates so all
crate::extension::test paths stay stable.

This collapses the test reverse-edge surface to 1 core file - mirroring the
bench (58->4) and trace (45->3) cluster collapses. The test phase subsystem is
now fully backed by contract types.
