+++ crates/homeboy-extension-contract/Cargo.toml
@@ -19,6 +19,7 @@ homeboy-product-identity = { version = "0.1.0", path = "../homeboy-product-ident
 homeboy-engine-primitives = { version = "0.1.0", path = "../homeboy-engine-primitives" }
 homeboy-lifecycle-contract = { version = "0.1.0", path = "../homeboy-lifecycle-contract" }
 homeboy-finding = { version = "0.1.0", path = "../homeboy-finding" }
+homeboy-refactor-contract = { version = "0.1.0", path = "../homeboy-refactor-contract" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 semver = "1.0"
